changed filter mask size

See merge request CaroloCup/freertoshal/fhal-driver!1

+++ include/can.h
@@ -42,7 +42,7 @@ struct can_filter {
 	/**
 	 * Mask
 	 */
-	uint16_t mask;
+	uint32_t mask;
 };
 #ifdef CONFIG_CAN_FD
 # define CAN_MAX_LENGTH 64
